% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_name_usage.R
\name{cb_name_usage}
\alias{cb_name_usage}
\title{Query ChecklistBank name usage}
\usage{
cb_name_usage(q = NULL, key = "3LXR", verbose = FALSE)
}
\arguments{
\item{q}{Query string (taxon name)}

\item{key}{ChecklistBank dataset key (default: "3LXR")}

\item{verbose}{Logical, return verbose output (default: FALSE)}
}
\value{
List with usage and alternatives data frames
}
\description{
Query the ChecklistBank API for taxonomic name usage information
}
