% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_name_usage_search.R
\name{cb_name_usage_search}
\alias{cb_name_usage_search}
\title{Search ChecklistBank by taxon ID}
\usage{
cb_name_usage_search(TAXON_ID = NULL, key = "3LXRC", limit = 1000)
}
\arguments{
\item{TAXON_ID}{Taxon ID to search for}

\item{key}{ChecklistBank dataset key (default: "3LXRC")}

\item{limit}{Maximum number of results (default: 1000)}
}
\value{
List with result data frame
}
\description{
Search the ChecklistBank API using a taxon ID
}
