% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{strip_html}
\alias{strip_html}
\title{Strip HTML tags from text}
\usage{
strip_html(html_text)
}
\arguments{
\item{html_text}{Character vector containing HTML}
}
\value{
Character vector with HTML tags removed
}
\description{
Helper function to remove HTML tags from labelHtml fields
}
