% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrong_rank.R
\name{wrong_rank}
\alias{wrong_rank}
\title{Check if a taxon has the wrong rank}
\usage{
wrong_rank(xx)
}
\arguments{
\item{xx}{A list containing name, wrongRank, and/or rightRank}
}
\value{
A character string indicating the issue status
}
\description{
Check if a taxon has the wrong rank
}
